requirements.txt
python-telegram-bot==13.11
Django==3.2.9
gunicorn==20.1.0
django-heroku==0.3.1
whitenoise==5.3.0
requests==2.27.1
